+++ tests/test_list_response.py
@@ -23,7 +23,7 @@ def test_user(load_sample):
         "schemas": ["urn:ietf:params:scim:api:messages:2.0:ListResponse"],
         "Resources": [resource_payload],
     }
-    response = ListResponse.of(User).model_validate(payload)
+    response = ListResponse[User].model_validate(payload)
     obj = response.resources[0]
     assert isinstance(obj, User)
 
@@ -37,7 +37,7 @@ def test_enterprise_user(load_sample):
         "schemas": ["urn:ietf:params:scim:api:messages:2.0:ListResponse"],
         "Resources": [resource_payload],
     }
-    response = ListResponse.of(User[EnterpriseUser]).model_validate(payload)
+    response = ListResponse[User[EnterpriseUser]].model_validate(payload)
     obj = response.resources[0]
     assert isinstance(obj, User[EnterpriseUser])
 
@@ -51,7 +51,7 @@ def test_group(load_sample):
         "schemas": ["urn:ietf:params:scim:api:messages:2.0:ListResponse"],
         "Resources": [resource_payload],
     }
-    response = ListResponse.of(Group).model_validate(payload)
+    response = ListResponse[Group].model_validate(payload)
     obj = response.resources[0]
     assert isinstance(obj, Group)
 
@@ -65,7 +65,7 @@ def test_service_provider_configuration(load_sample):
         "schemas": ["urn:ietf:params:scim:api:messages:2.0:ListResponse"],
         "Resources": [resource_payload],
     }
-    response = ListResponse.of(ServiceProviderConfig).model_validate(payload)
+    response = ListResponse[ServiceProviderConfig].model_validate(payload)
     obj = response.resources[0]
     assert isinstance(obj, ServiceProviderConfig)
 
@@ -82,7 +82,7 @@ def test_resource_type(load_sample):
         "schemas": ["urn:ietf:params:scim:api:messages:2.0:ListResponse"],
         "Resources": [user_resource_type_payload, group_resource_type_payload],
     }
-    response = ListResponse.of(ResourceType).model_validate(payload)
+    response = ListResponse[ResourceType].model_validate(payload)
     obj = response.resources[0]
     assert isinstance(obj, ResourceType)
 
@@ -100,7 +100,7 @@ def test_mixed_types(load_sample):
         "schemas": ["urn:ietf:params:scim:api:messages:2.0:ListResponse"],
         "Resources": [user_payload, group_payload],
     }
-    response = ListResponse.of(Union[User, Group]).model_validate(payload)
+    response = ListResponse[Union[User, Group]].model_validate(payload)
     user, group = response.resources
     assert isinstance(user, User)
     assert isinstance(group, Group)
@@ -124,16 +124,16 @@ def test_mixed_types_type_missing(load_sample):
         "Resources": [user_payload, group_payload],
     }
 
-    ListResponse.of(User, Group).model_validate(payload)
+    ListResponse[Union[User, Group]].model_validate(payload)
 
     with pytest.raises(ValidationError):
-        ListResponse.of(User, Foobar).model_validate(payload)
+        ListResponse[Union[User, Foobar]].model_validate(payload)
 
     with pytest.raises(ValidationError):
-        ListResponse.of(User).model_validate(payload)
+        ListResponse[User].model_validate(payload)
 
 
-def test_missing_resource_schema(load_sample):
+def test_missing_resource_payload(load_sample):
     """Check that validation fails if resources schemas are missing."""
 
     payload = {
@@ -145,10 +145,28 @@ def test_missing_resource_schema(load_sample):
     }
 
     with pytest.raises(ValidationError):
-        ListResponse.of(User, Group).model_validate(payload, strict=True)
+        ListResponse[Union[User, Group]].model_validate(payload, strict=True)
 
     # TODO: This should raise a ValidationError
-    ListResponse.of(User).model_validate(payload, strict=True)
+    ListResponse[User].model_validate(payload, strict=True)
+
+
+def test_missing_resource_schema(load_sample):
+    """Check that validation fails if resources schemas are missing."""
+
+    payload = {
+        "totalResults": 2,
+        "itemsPerPage": 10,
+        "startIndex": 1,
+        "schemas": ["urn:ietf:params:scim:api:messages:2.0:ListResponse"],
+        "Resources": [{"id": "foobar"}],
+    }
+
+    with pytest.raises(ValidationError):
+        ListResponse[Union[User, Group]].model_validate(payload, strict=True)
+
+    # TODO: This should raise a ValidationError
+    ListResponse[User].model_validate(payload, strict=True)
 
 
 def test_zero_results():
@@ -171,19 +189,17 @@ def test_zero_results():
             }
         ],
     }
-    ListResponse.of(User).model_validate(
-        payload, scim_ctx=Context.RESOURCE_QUERY_RESPONSE
-    )
+    ListResponse[User].model_validate(payload, scim_ctx=Context.RESOURCE_QUERY_RESPONSE)
 
     payload = {"totalResults": 1, "Resources": []}
     with pytest.raises(ValidationError):
-        ListResponse.of(User).model_validate(
+        ListResponse[User].model_validate(
             payload, scim_ctx=Context.RESOURCE_QUERY_RESPONSE
         )
 
     payload = {"totalResults": 1}
     with pytest.raises(ValidationError):
-        ListResponse.of(User).model_validate(
+        ListResponse[User].model_validate(
             payload, scim_ctx=Context.RESOURCE_QUERY_RESPONSE
         )
 
@@ -204,4 +220,4 @@ def test_list_response_schema_ordering():
             }
         ],
     }
-    ListResponse.of(User[EnterpriseUser], Group).model_validate(payload)
+    ListResponse[Union[User[EnterpriseUser], Group]].model_validate(payload)